Move EsqlVersion to core
This enables using the version from outside the esql project, esp. in
integration tests.
