+++ x-pack/plugin/core/src/main/java/module-info.java
@@ -67,6 +67,7 @@ module org.elasticsearch.xcore {
     exports org.elasticsearch.xpack.core.esql;
     exports org.elasticsearch.xpack.core.esql.action;
     exports org.elasticsearch.xpack.core.esql.action.internal; // TODO: qualify to esql when modularized
+    exports org.elasticsearch.xpack.core.esql.version;
     exports org.elasticsearch.xpack.core.frozen.action;
     exports org.elasticsearch.xpack.core.frozen;
     exports org.elasticsearch.xpack.core.graph.action;
