+++ x-pack/plugin/esql/qa/server/build.gradle
@@ -1,4 +1,5 @@
 apply plugin: 'elasticsearch.java'
+apply plugin: 'elasticsearch.internal-es-plugin'
 
 description = 'Integration tests for ESQL'
 
@@ -9,5 +10,6 @@ dependencies {
   api project(xpackModule('ql:test-fixtures'))
   // Requirement for some ESQL-specific utilities
   implementation project(':x-pack:plugin:esql')
+  implementation project(path: xpackModule('core'))
   api project(xpackModule('esql:qa:testFixtures'))
 }
